halfway fix for the gl buffer/texture delete problem
with this change the bsp/mdl loaded by the server no longer create gl
objects.

while this is not a real fix, as we can still not delete the gl objects
it allows to change levels again. the models just get never deleted but
cached inside the map 'models' of modelconf.go